split up thread lib
