+++ .clang-format
@@ -49,7 +49,7 @@ BasedOnStyle: Microsoft
 #   - __capability
 # BinPackArguments: true
 # BinPackParameters: true
-# BraceWrapping:
+BraceWrapping:
 #   AfterCaseLabel:  true
 #   AfterClass:      true
 #   AfterControlStatement: Always
@@ -58,7 +58,7 @@ BasedOnStyle: Microsoft
 #   AfterNamespace:  true
 #   AfterObjCDeclaration: true
 #   AfterStruct:     true
-#   AfterUnion:      true
+   AfterUnion:      true
 #   AfterExternBlock: false #
 #   BeforeCatch:     true
 #   BeforeElse:      true
